Set account roles from $roles
