+++ src/routes/Tokens.svelte
@@ -4,7 +4,7 @@
         accountRoles,
         activeNetwork,
         auditHistory,
-        ethersData,
+        ethersData, roles,
         schemas,
         tokens,
         vault
@@ -32,7 +32,7 @@
             localStorage.setItem("vaultAddress", token.address)
             let auditHistoryData = await getAuditHistoryData(token.address)
             auditHistory.set(auditHistoryData)
-            accountRoles.set(await setAccountRoles($vault, $activeNetwork, $account));
+            accountRoles.set(await setAccountRoles($roles, $account));
             navigateTo("#roles")
         }
     }
